+++ src/inspector/types/credentials/context.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod';
+
+export const ContextSchema = z.object({
+  name: z.string().optional(),
+  url: z.string().url(),
+});
+
+export type Context = {
+  name?: string;
+  url: URL;
+};
