+++ src/inspector/types/credentials/credential.ts
@@ -1,0 +1,4 @@
+export type CredentialSchema = {
+  type: string;
+  id: URL;
+};
