+++ src/inspector/types/credentials/eIDAS.ts
@@ -1,0 +1,3 @@
+export type eIDAS2 = {
+  validToStandard: boolean;
+};
