+++ src/inspector/types/credentials/format.ts
@@ -1,0 +1,8 @@
+export type Format = {
+  name: string;
+  /*
+   * If the format is not JSON,
+   * this value is the JSON representation
+   */
+  convertedToJSON?: string;
+};
