+++ src/inspector/types/credentials/proof.ts
@@ -1,0 +1,3 @@
+export type Proof = {
+  type: SecuringMechanism;
+};
