+++ src/inspector/types/credentials/securingMechanism.ts
@@ -1,0 +1,1 @@
+export type SecuringMechanism = 'Enveloping' | 'Embedded';
