+++ src/inspector/types/credentials/subject.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export type Subject = {
+  id?: URL;
+  claims: Claim[];
+};
