+++ src/inspector/types/credentials/validity.ts
@@ -1,0 +1,6 @@
+export type Validity = {
+  isValid: boolean;
+  validFrom?: Date;
+  validUntil?: Date;
+  proof: ListResult<Proof>;
+};
