+++ src/inspector/types/credentials/verifiablePresentation.ts
@@ -1,0 +1,9 @@
+/**
+ * Verifiable Presentation
+ */
+export type VP = {
+  type: Result<string[]>;
+  context: Result<Context[]>;
+  credentials: Result<VC>[];
+  proofs: ListResult<Proof>;
+};
